added navBar && header fixed && styled with chakra
